pyproject.toml
[project]
name = "langchain-course-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.9.0",
    "boto3>=1.40.42",
    "faiss-cpu>=1.12.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.34",
    "langchain-community>=0.3.30",
    "langchainhub>=0.1.21",
    "opensearch-py>=3.0.0",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "requests-aws4auth>=1.3.1",
]
